Fixing pyenv shell integration and some macOS stuff

+++ shell/common/source/mac/functions/functions_mac.sh
@@ -11,7 +11,7 @@ prepare_mac() {
 # before any useful software development can be performed.
 install_mac_developer_tools() {
   # Make sure developer tools are installed
-  xcode-select --install >>/dev/null || :
+  xcode-select --install >>/dev/null 2>&1 || :
 }
 
 # Install the homebrew package manager for macOS.
@@ -25,6 +25,12 @@ install_homebrew() {
 # Install all homebrew packages
 install_mac_packages() {
   local package
+
+  if [ ! -f "${BREW_PACKAGES_FILE}" ]; then
+    err "No homebrew packages file found at: ${BREW_PACKAGES_FILE}"
+    return 1
+  fi
+
   while read -r package || [ -n "${package}" ]; do
     brew install "${package}"
   done <"${BREW_PACKAGES_FILE}"
@@ -37,20 +43,21 @@ create_gnu_cli_tool_aliases_for_mac() {
   local brewPrefix="$(brew --prefix)/opt"
   local pathToGnubin="libexec/gnubin"
 
-  __add_tool_to_path "${brewPrefix}/binutils/bin"
-  __add_tool_to_path "${brewPrefix}/coreutils/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/findutils/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/gawk/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/gnu-sed/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/gnu-getopt/bin"
-  __add_tool_to_path "${brewPrefix}/gnu-indent/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/gnu-tar/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/gnu-which/${pathToGnubin}"
-  __add_tool_to_path "${brewPrefix}/grep/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/binutils/bin"
+  __add_mac_tool_to_path "${brewPrefix}/coreutils/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/findutils/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/gawk/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/gnu-sed/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/gnu-getopt/bin"
+  __add_mac_tool_to_path "${brewPrefix}/gnu-indent/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/gnu-tar/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/gnu-which/${pathToGnubin}"
+  __add_mac_tool_to_path "${brewPrefix}/grep/${pathToGnubin}"
 }
 
-# Add a given path to the $PATH environment variable.
-__add_tool_to_path() {
+# Add a given tool's path to the $PATH environment variable if and only if it is
+# not present on $PATH already.
+__add_mac_tool_to_path() {
   local pathToToolBinDir="${1}"
   if ! echo "${PATH}" | grep -q "${pathToToolBinDir}" && \
     ([ -d "${pathToToolBinDir}" ] || [ -h "${pathToToolBinDir}" ]); then
